Class management for optimization, a few changes in diagram class including backend updates.
