+++ backend/WatchNow/src/main/java/com/intiformation/WatchNow/model/Oeuvre.java
@@ -2,16 +2,15 @@ package com.intiformation.WatchNow.model;
 
 
 import java.time.LocalDate;
-import java.util.*;
 
 public class Oeuvre {
 
     protected String id;
-    protected Title title;
-    protected Ratings ratings;
+    protected OeuvreTitle title;
+    protected OeuvreRatings ratings;
     protected String[] genres;
     protected LocalDate releaseDate;
-    protected PlotSummary plotSummary;
+    protected OeuvrePlotSummary plotSummary;
     protected Synopsis synopsis;
     protected Trailer trailer;
 
@@ -54,11 +53,11 @@ public class Oeuvre {
 		this.releaseDate = releaseDate;
 	}
 
-    public String getTitle() {
+    public String getTitre() {
     	return title.getTitle();
     }
     
-    public void setTitle(String title) {
+    public void setTitre(String title) {
     	this.title.setTitle(title);
     }
 
@@ -126,164 +125,11 @@ public class Oeuvre {
 		this.trailer = trailer;
 	}
 
-	public String getImageUrl() {
-		return this.title.getImage().getUrl();
+	public OeuvreImage getImage() {
+		return this.title.getImage();
 	}
 
-	public void setImageUrl(String url) {
-		this.title.getImage().setUrl(url);
+	public void setImageUrl(OeuvreImage image) {
+		this.title.setImage(image);
 	}
-	
-    
-    
-}
-
-class Title {
-	private String id;
-	private String title;
-	private String titleType;
-	private Integer runningTimeInMinutes;
-	private Image image;
-	private Integer numberOfEpisodes;
-	private Integer seriesStartYear;
-	private Integer seriesEndYear;
-	private Integer episode;
-	private Integer season;
-	private String nextEpisode;
-	private String previousEpisode;
-	private Title parentTitle;
-	
-	public String getId() {
-		return id.split("/")[id.split("/").length-1];
-	}
-	public void setId(String id) {
-		this.id = id;
-	}
-	public String getTitle() {
-		return title;
-	}
-	public void setTitle(String title) {
-		this.title = title;
-	}
-	public String getTitleType() {
-		return titleType;
-	}
-	public void setTitleType(String titleType) {
-		this.titleType = titleType;
-	}
-	public Integer getEpisode() {
-		return episode;
-	}
-	public void setEpisode(Integer episode) {
-		this.episode = episode;
-	}
-	public Integer getSeason() {
-		return season;
-	}
-	public void setSeason(Integer season) {
-		this.season = season;
-	}
-	public Integer getRunningTimeInMinutes() {
-		return runningTimeInMinutes;
-	}
-	public void setRunningTimeInMinutes(Integer runningTimeInMinutes) {
-		this.runningTimeInMinutes = runningTimeInMinutes;
-	}
-	public String getNextEpisode() {
-		return nextEpisode.split("/")[id.split("/").length-1];
-	}
-	public void setNextEpisode(String nextEpisode) {
-		this.nextEpisode = nextEpisode;
-	}
-	public String getPreviousEpisode() {
-		return previousEpisode.split("/")[id.split("/").length-1];
-	}
-	public void setPreviousEpisode(String previousEpisode) {
-		this.previousEpisode = previousEpisode;
-	}
-	public Integer getNumberOfEpisodes() {
-		return numberOfEpisodes;
-	}
-	public void setNumberOfEpisodes(Integer numberOfEpisodes) {
-		this.numberOfEpisodes = numberOfEpisodes;
-	}
-	public Integer getSeriesStartYear() {
-		return seriesStartYear;
-	}
-	public void setSeriesStartYear(Integer seriesStartYear) {
-		this.seriesStartYear = seriesStartYear;
-	}
-	public Integer getSeriesEndYear() {
-		return seriesEndYear;
-	}
-	public void setSeriesEndYear(Integer seriesEndYear) {
-		this.seriesEndYear = seriesEndYear;
-	}
-	public Image getImage() {
-		return image;
-	}
-	public void setImage(Image image) {
-		this.image = image;
-	}
-	public Title getParentTitle() {
-		return parentTitle;
-	}
-	public void setParentTitle(Title parentTitle) {
-		this.parentTitle = parentTitle;
-	}
-	
-}
-
-class Ratings {
-	private Double rating;
-	private Integer ratingCount;
-	public Double getRating() {
-		return rating;
-	}
-	public void setRating(Double rating) {
-		this.rating = rating;
-	}
-	public Integer getRatingCount() {
-		return ratingCount;
-	}
-	public void setRatingCount(Integer ratingCount) {
-		this.ratingCount = ratingCount;
-	}
-}
-
-class PlotSummary {
-	private String author;
-	private String text;
-	public String getAuthor() {
-		return author;
-	}
-	public void setAuthor(String author) {
-		this.author = author;
-	}
-	public String getText() {
-		return text;
-	}
-	public void setText(String text) {
-		this.text = text;
-	}
-	
-}
-
-class Image{
-	private String id;
-	private String url;
-	public String getId() {
-		return id;
-	}
-	public void setId(String id) {
-		this.id = id;
-	}
-	public String getUrl() {
-		return url;
-	}
-	public void setUrl(String url) {
-		this.url = url;
-	}
-	
-	
 }
