+++ backend/WatchNow/src/main/java/com/intiformation/WatchNow/model/OeuvreImage.java
@@ -1,0 +1,19 @@
+package com.intiformation.WatchNow.model;
+
+public class OeuvreImage {
+	private String id;
+	private String url;
+	public String getId() {
+		return id;
+	}
+	public void setId(String id) {
+		this.id = id;
+	}
+	public String getUrl() {
+		return url;
+	}
+	public void setUrl(String url) {
+		this.url = url;
+	}
+
+}
