+++ backend/WatchNow/src/main/java/com/intiformation/WatchNow/model/OeuvrePlotSummary.java
@@ -1,0 +1,19 @@
+package com.intiformation.WatchNow.model;
+
+public class OeuvrePlotSummary {
+
+	private String author;
+	private String text;
+	public String getAuthor() {
+		return author;
+	}
+	public void setAuthor(String author) {
+		this.author = author;
+	}
+	public String getText() {
+		return text;
+	}
+	public void setText(String text) {
+		this.text = text;
+	}
+}
