+++ backend/WatchNow/src/main/java/com/intiformation/WatchNow/model/OeuvreRatings.java
@@ -1,0 +1,19 @@
+package com.intiformation.WatchNow.model;
+
+public class OeuvreRatings {
+
+	private Double rating;
+	private Integer ratingCount;
+	public Double getRating() {
+		return rating;
+	}
+	public void setRating(Double rating) {
+		this.rating = rating;
+	}
+	public Integer getRatingCount() {
+		return ratingCount;
+	}
+	public void setRatingCount(Integer ratingCount) {
+		this.ratingCount = ratingCount;
+	}
+}
